+++ core/src/main/java/io/cucumber/core/cli/CommandlineOptions.java
@@ -47,11 +47,6 @@ public final class CommandlineOptions {
     public static final String DRY_RUN = "--dry-run";
     public static final String DRY_RUN_SHORT = "-d";
 
-    public static final String NO_STRICT = "--no-strict";
-
-    public static final String STRICT = "--strict";
-    public static final String STRICT_SHORT = "-s";
-
     public static final String NO_MONOCHROME = "--no-monochrome";
 
     public static final String MONOCHROME = "--monochrome";